changing types in first function

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -6,10 +6,10 @@
  * Return: count
  */
 
-int countSetBits(int n)
+int countSetBits(unsigned long int n)
 {
 
-	int count = 0;
+	unsigned long int count = 0;
 
 	if (n <= 0)
 		return (0);
